give e2e common structure
