+++ integration/setup/setup.go
@@ -1,0 +1,32 @@
+// +build k8srequired
+
+package setup
+
+import (
+	"log"
+	"os"
+	"testing"
+
+	"github.com/giantswarm/e2e-harness/pkg/framework"
+)
+
+func WrapTestMain(f *framework.Host, m *testing.M) {
+	var v int
+	var err error
+
+	err = f.CreateNamespace("giantswarm")
+	if err != nil {
+		log.Printf("%#v\n", err)
+		v = 1
+	}
+
+	if v == 0 {
+		v = m.Run()
+	}
+
+	if os.Getenv("KEEP_RESOURCES") != "true" {
+		f.Teardown()
+	}
+
+	os.Exit(v)
+}
