+++ integration/test/basic/fixtures/test-chart/templates/pod.yaml
